+++ contracts/interfaces/ITimelockControllerOptimistic.sol
@@ -4,11 +4,20 @@ pragma solidity ^0.8.28;
 interface ITimelockControllerOptimistic {
     error TimelockControllerOptimistic__OperationConflict();
     error TimelockControllerOptimistic__UnauthorizedUpgrade();
+    error TimelockControllerOptimistic__InvalidVersionRegistry();
+    error TimelockControllerOptimistic__VersionRegistryAlreadySet();
 
     // === Events ===
 
-    function initialize(uint256 minDelay, address[] memory proposers, address[] memory executors, address admin)
-        external;
+    event VersionRegistrySet(address versionRegistry);
+
+    function initialize(
+        uint256 minDelay,
+        address[] memory proposers,
+        address[] memory executors,
+        address admin,
+        address _versionRegistry
+    ) external;
 
     function executeBatchBypass(
         address[] calldata targets,